% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyBfastmonitor.R
\name{plotlyBfm}
\alias{plotlyBfm}
\title{Converts the plot.bfastmonitor function to a plotly graph}
\usage{
plotlyBfm(bfm)
}
\arguments{
\item{bfm}{a bfastmonitor object.}
}
\description{
Converts the plot.bfastmonitor function to a plotly graph
}
\examples{
require(bfast)
require(zoo)
NDVIa <- as.ts(zoo(som$NDVI.a, som$Time))
mona <- bfastmonitor(NDVIa, start = c(2010, 13))
plotlyBfm(mona)

}
