% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyBfastmonitor.R
\name{num2time}
\alias{num2time}
\title{Converts a numeric time to year-cycle representation}
\usage{
num2time(bfm)
}
\arguments{
\item{bfm}{a bfastmonitor object.}
}
\description{
Converts a numeric time to year-cycle representation, i.e. 2011(3)
}
\examples{
require(bfast)
require(plotlyBFAST)
NDVIa <- as.ts(zoo(som$NDVI.a, som$Time))
mona <- bfastmonitor(NDVIa, start = c(2010, 13))
num2time(mona)

}
